Changes made to the list length

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -7,7 +7,7 @@
  */
 size_t list_len(const list_t *h)
 {
-size_t num_nodes;
+unsigned short int num_nodes;
 while (h != NULL)
 {
 num_nodes += 1;
